D2: override with same return type (ITimeoutManager)
D1 wrongly allows to override a method with a return type of different
integer precision

+++ src/ocean/time/timeout/model/ITimeoutManager.d
@@ -80,7 +80,7 @@ interface ITimeoutManager
 
     ***************************************************************************/
 
-    uint checkTimeouts ( bool delegate ( ITimeoutClient client ) dg );
+    size_t checkTimeouts ( bool delegate ( ITimeoutClient client ) dg );
 
     /***************************************************************************
